Release v24.2

+++ daemon/frontend/Frontend.h
@@ -3,6 +3,7 @@
  *
  *  Copyright (C) 2004 Sven Henkel <sidddy@users.sourceforge.net>
  *  Copyright (C) 2007-2016 Andrey Prygunkov <hugbug@users.sourceforge.net>
+ *  Copyright (C) 2024 Denis <denis@nzbget.com>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -15,7 +16,7 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 
@@ -42,7 +43,7 @@ protected:
 	int m_updateInterval;
 
 	// summary
-	int m_currentDownloadSpeed = 0;
+	int64 m_currentDownloadSpeed = 0;
 	int64 m_remainingSize = 0;
 	bool m_pauseDownload = false;
 	int m_downloadLimit = 0;
