+++ daemon/main/Scheduler.cpp
@@ -2,6 +2,7 @@
  *  This file is part of nzbget. See <https://nzbget.com>.
  *
  *  Copyright (C) 2008-2019 Andrey Prygunkov <hugbug@users.sourceforge.net>
+ *  Copyright (C) 2024 Denis <denis@nzbget.com>
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -14,7 +15,7 @@
  *  GNU General Public License for more details.
  *
  *  You should have received a copy of the GNU General Public License
- *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 
@@ -65,7 +66,7 @@ void Scheduler::ScheduleNextWork()
 	sched.tm_sec = 0;
 	time_t nextMinute = Util::Timegm(&sched);
 
-	m_serviceInterval = nextMinute - curTime;
+	m_serviceInterval = static_cast<int>(nextMinute - curTime);
 }
 
 void Scheduler::ServiceWork()
